Repositioning and Rotating 45 degrees
Changed the squares so that the least significant bit is in the lower
left square and the most significant bit is in the upper right. So the
lower left is index 0 in the binary number and the upper right is index
63.

I fixed the code where necessary for this change.

Added functions to pseudo-rotate the bitboard 45 degrees CW and CCW.
This will be used for diagonal sliding pieces.

Added a new constructor that creates a chessboard from a list of longs
that define where each type of piece is.

Need more test cases for pawn movements.
Need test cases for king,rook, and knight moves.
Need to do diagonal sliding pieces to finish bishop and queen and to add
test cases.