Sliding Moves
Completely changed the horizontal sliding method to match the diagonal
sliding method. Now it maps the square position to a hashmap that maps
occupancy to moves. Before it mapped file to the hashmap.

Fixed the diagonal sliding method.

Minor adjustments to the rook move generation and queen move generation
to work with the updated horizontal sliding method.