Redesigned the Principal Variation search and transposition tables
remade the PV search algorithm so that it works from the actual root
node. Saw 2-4x increase in speed. this is now the normal

transposition tables have been implemented. Saw a 2x increase in speed
compared to normal (with the updated PV search algorithm). This was
primaily visible when analyzing a board for a second time or responding
to a move. Very useful when getting to depths of 5+.

next on the todo list:  --> import a PGN to a chess game so that I can
analyze the game using regression.